add stripe payment after upload
