+++ experiences/vuetify.options.js
@@ -13,6 +13,7 @@ import {
   mdiChevronLeft,
   mdiChevronRight,
   mdiCircleOutline,
+  mdiCheckCircleOutline,
   mdiCircleSlice8,
   mdiClose,
   mdiCloseBox,
@@ -67,6 +68,7 @@ const values = {
   mdiChartBar,
   mdiCheckboxBlankOutline,
   mdiCheckCircle,
+  mdiCheckCircleOutline,
   mdiChevronLeft,
   mdiChevronRight,
   mdiCircleOutline,
